fix permission system

+++ commands/music/resume.ts
@@ -28,7 +28,8 @@ export default class Resume extends Command {
         const message = this.message
         const embeds = new Embeds(discord, message)
         const audio = new Audio(discord, message)
-if (!audio.checkMusicPlaying()) return
+        if (!audio.checkMusicPermissions()) return
+        if (!audio.checkMusicPlaying()) return
         audio.resume()
         const rep = await message.reply("Resumed the song!")
         rep.delete({timeout: 3000}).then(() => message.delete().catch(() => null))
